[DEV-14238] - WIP transaction flow refactor
